Several Changes - Incremented Dev Version 0.0.2
- Modified router/default.py to register any consumer based on the
router "connect" calls
- Modified router/default.py to register and destination queue based on
router "connect" calls in an array
- Added "send_event" method on consumer.py to send a copied even to all
"outbox" queues connected to the module

This is a part of the overall direction shift in the inherent behavior.
All modules will accept input from anyone, if configured to do so with
router.connect(), and all modules will, by default, replicate events to
ALL outbox queues, as configured via router.connect().

In other words, every module inherently supports a N to N relationship.

This needs to be refined and enhanced - but this is the initial commit
for those underlying direction changes