Changed Logging and added flowcontroller
Added FlowController module, which is a simple module designed to
abstract complicated data flow down to one (or more as the case may be)
queues. The explicit and immediate purpose of this module was to
abstract multiple data flows to an XMLFilter.

Also, almost all instances of print() were changed to their respective
self.logging severity level excecutions.