Fixed get_actor for internal use actors (logs etc)
get_actor broke when invoked against a logging, metrics, or failed
actors. Also, added a temporary behavior change: metrics are no longer
emitted by any actor if a metric actor is "Null" at the time of
Director.start() call. It's just wasted processing at that point.

In the future, "Null" actors will not be the default behavior, but for
now this patch will suffice